[client-app] better handling of startup errors
Summary:
We ALWAYS popped open the damn inspector panel if there was a startup
error in the browser window. This would happen when we got EPERM issues
from the compile cache :(

Reporting from here is tricky since the browser windows don't setup their
error reporter until the NylasEnv.initialize call. Instead we send it to
the backend main process to handle and report.

Since there are multiple windows booting up at once, I needed only 1 of
them to pop up the dialog to users. Calling app.quit() from all 3 windows
in a row with 3 dialog boxes caused the app to not exist since the dialog
boxes stops proper code execution. This would leave the backend process
dead and the windows as zombies forever :(

This can be fixed by only having 1 window show the error message. I chose
just the "work" window since this error tends to happen to all windows and
it's the window most likely to be started on launch.

Test Plan:
Manually threw an error in the index.js and asserted that the error
reporter in the backend main process properly received it and emitted.
Also ensured that there was only 1 dialog box and that it fully quit the
app with all windows.

Reviewers: halla, juan

Reviewed By: juan

Subscribers: gleb

Differential Revision: https://phab.nylas.com/D4453