[*] Fix the specs
Summary:
Turns out that there was a spec for the IdentityStore that made the
spec runner exit immediately without an error code. I'm not sure what
exactly causes that, so for now I've disabled the spec. This diff
also fixes some specs that have gotten out-of-date since then.

Test Plan: Ran the specs

Reviewers: juan, evan, mark

Reviewed By: evan, mark

Differential Revision: https://phab.nylas.com/D4474