[client-app] Fix tracking when sending per recipient
Summary:
When sending custom bodies per recipient for tracking, we used to copy
the message model, and then update the copy's body with the custom
body that contains the tracking links for each recipient.

However, due to recent changes that store message bodies in flat files on
disk, copying an instance of a message would actually update the body of
the original message because they both pointed to the same message body
file on disk, which is obviously not what this code initially intended
to do.

This caused all messages to go out with the tracking links of the first
recipient, causing open tracking to misreport who had actually opened a
message.

The fix for now is to not make a copy of the message model, but rather
just of its values, which we can use for sending.

Test Plan: manually test that send-per-recipient and tracking works as expected

Reviewers: evan, halla, mark

Reviewed By: halla, mark

Differential Revision: https://phab.nylas.com/D4496